DDD Updated with more servers attributes Simple Service Updated

+++ Services/DDD/Inventory.Domain/Models/ServerGroup.cs
@@ -7,10 +7,10 @@ namespace Inventory.Domain.Models
 {
     public class ServerGroup
     {
-        public Int64 ServerId { get; set; }
+        public int ServerId { get; set; }
         public Server Server { get; set; }
 
-        public Int64 GroupId { get; set; }
+        public int GroupId { get; set; }
         public Group Group { get; set; }
 
     }
